refactor : Label을 코드 응집성을 고려해 User package 안으로 이동
