--- src/main/java/io/github/eappezo/soundary/services/label/api/dto/LabelDeleteRequest.java
@@ -1,7 +1,0 @@
-package io.github.eappezo.soundary.services.label.api.dto;
-
-public record LabelDeleteRequest(
-    String label
-) {
-
-}
