+++ src/main/java/io/github/eappezo/soundary/services/user/api/controller/LabelController.java
@@ -1,16 +1,16 @@
-package io.github.eappezo.soundary.services.label.api.controller;
+package io.github.eappezo.soundary.services.user.api.controller;
 
 import io.github.eappezo.soundary.core.identification.Identifier;
-import io.github.eappezo.soundary.services.label.api.dto.LabelAddRequest;
-import io.github.eappezo.soundary.services.label.api.dto.LabelDeleteRequest;
-import io.github.eappezo.soundary.services.label.api.dto.LabelListResponse;
-import io.github.eappezo.soundary.services.label.application.dto.UserLabelDto;
-import io.github.eappezo.soundary.services.label.application.service.LabelService;
+import io.github.eappezo.soundary.services.user.api.dto.LabelAddRequest;
+import io.github.eappezo.soundary.services.user.api.dto.LabelListResponse;
+import io.github.eappezo.soundary.services.user.application.dto.UserLabelDto;
+import io.github.eappezo.soundary.services.user.application.service.LabelService;
 import java.net.URI;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
@@ -30,10 +30,10 @@ public class LabelController implements LabelAPI {
     }
 
     @Override
-    @DeleteMapping()
+    @DeleteMapping("/{label}")
     public ResponseEntity<Void> deleteLabel(Identifier userId,
-        LabelDeleteRequest labelDeleteRequest) {
-        labelService.deleteLabel(userId, labelDeleteRequest.label());
+        @PathVariable String label) {
+        labelService.deleteLabel(userId, label);
         return ResponseEntity.noContent().build();
     }
