+++ src/main/java/io/github/eappezo/soundary/services/user/infrastructure/persistence/entity/UserLabelEntity.java
@@ -1,6 +1,8 @@
-package io.github.eappezo.soundary.services.label.domain;
+package io.github.eappezo.soundary.services.user.infrastructure.persistence.entity;
 
 import io.github.eappezo.soundary.core.persistence.infrastructure.BaseEntity;
+import io.github.eappezo.soundary.services.user.domain.Label;
+import io.github.eappezo.soundary.services.user.infrastructure.persistence.entity.key.UserLabelEntityKey;
 import jakarta.persistence.Entity;
 import jakarta.persistence.EnumType;
 import jakarta.persistence.Enumerated;
