+++ src/main/java/io/github/eappezo/soundary/services/user/infrastructure/persistence/entity/key/UserLabelEntityKey.java
@@ -1,5 +1,6 @@
-package io.github.eappezo.soundary.services.label.domain;
+package io.github.eappezo.soundary.services.user.infrastructure.persistence.entity.key;
 
+import io.github.eappezo.soundary.services.user.domain.Label;
 import java.io.Serializable;
 import lombok.AllArgsConstructor;
 import lombok.EqualsAndHashCode;
